Liveness clock progress (#163)
* Prove eventually tick theorem for simple liveness protocol

This finishes the eventually ticks theorem for the
first example protocol for liveness.
This makes a necessary change to the defintion of the
validator VLSM to make it wait for all plan-expected
messages before ending a timestep.

The rest of the work is in proofs, defining and proving
the invariants necessary to complete the proof that the
variant decreases, and adding the main theorem that
applies the variant to prove that clocks eventually tick.

Several of the definitions and lemmas needed to
finish the clock advance proof were not specific
to this particular proof and protocol, or could be
obviously generalized. These have been added in
appropriate files outside of VLSM.Liveness.

The biggest change to existing proofs was simplifying
the composite_has_been_sent_capability instance
in the process of adding the new instance
composite_has_been_observed_capability.